Split what used to be “packer” into multiple utils
This will make it waaay easier to unit-test this suckr!
