+++ app/src/shared/components/PermissionsProvider/utils/lookup.js
@@ -1,0 +1,6 @@
+import * as operations from '../operations'
+
+export default function lookup(combination, operationKey) {
+    // eslint-disable-next-line no-bitwise
+    return !!(combination & operations[operationKey])
+}
