+++ app/src/shared/components/PermissionsProvider/utils/toOperationId.js
@@ -1,0 +1,6 @@
+import * as operations from '../operations'
+import toOperationKey from './toOperationKey'
+
+export default function toOperationID(o) {
+    return operations[toOperationKey(o)]
+}
