+++ app/src/shared/components/PermissionsProvider/utils/toOperationKey.js
@@ -1,0 +1,3 @@
+export default function toOperationKey(o) {
+    return o.toUpperCase().split(/_/).reverse()[0]
+}
