+++ app/src/shared/components/PermissionsProvider/utils/toOperationName.js
@@ -1,0 +1,5 @@
+import toOperationKey from './toOperationKey'
+
+export default function toOperationName(resourceType, o) {
+    return `${resourceType}_${toOperationKey(o)}`.toLowerCase()
+}
